12. Commit - 22.09.2023
validate_copied_text() is added. In summary, this code is a function that checks whether certain conditions are met in a given text. It counts the number of tabs (tabs are converted to lowercase), converts the entire text to lowercase, and then checks if any of the specified words ("adet," "kg," "pk," "mt," or "metre") are present in the lowercase text. If there are more than 2 tabs and at least one of these words is found, it returns True; otherwise, it returns False.